+++ src/books/routes.js
@@ -2,12 +2,18 @@ const { Router } = require("express");
 const bookRouter = Router();
 const Book = require("./model");
 
-const { addBook, findBooks } = require(".controllers");
+const { addBook } = require("./controllers");
+const { findBook } = require("./controllers");
+const { getFirstBook } = require("./controllers");
+const { updateAuthor } = require("./controllers");
 
 //Create a new book entry in the database using POST request function.
 //##############################################
 bookRouter.post("/books/addbook", addBook);
-bookRouter.get("/books/findBooks", findBooks);
+bookRouter.get("/books/findBooks", findBook);
+bookRouter.get("/books/getFirstBook", getFirstBook);
+bookRouter.put("/books/updateAuthor", updateAuthor);
+
 //###############################################
 //anatomy of a route
